Community code merge (#223)
* MNT-25021 (#691)

Earlier, we first fetched the app config and then checked the auth type. Now, we've changed the approach: we first check the auth type, and based on that, decide whether to fetch the app config. If the auth type is 'basic', we skip fetching the app config.

* Revert "MNT-25021 (#691)" (#692)

This reverts commit 2b803cf4d8a9e57f2d7a02ae5065cc2f17577451.

* Mnt 25021 (#693)

* MNT-25021

Earlier, we first fetched the app config and then checked the auth type. Now, we've changed the approach: we first check the auth type, and based on that, decide whether to fetch the app config. If the auth type is 'basic', we skip fetching the app config.

* Updated Code

* Added response code

* Used Enum

* Pod Updated (#694)

AlfrescoCore Updated